pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doe-metrics-reporter"
version = "0.1.0"
description = "AI-powered metrics reporter for NERSC system health using MCP architecture"
requires-python = ">=3.11"
authors = [
    {name = "NERSC User", email = "user@nersc.gov"}
]
dependencies = [
    "httpx>=0.28.1",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.100.0",
    "slack-sdk>=3.27.0",
    "mcp>=1.0.0",
    "fastmcp>=0.1.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
